subject.h: update `detach` to require a pointer to an `Observer`
Since `ChangeManager` is expected to act as a mediator between `Subject`
and `Observer`, it is responsible for notifying the right `Observer`(s).

The `ChangeManager` uses an unordered_multimap for tracking the
`Subject`-`Observer` relationships. Thus, modify the arguments passed to
the mediator to be a pointer to the `Observer` to be detached instead.

Signed-off-by: Anant Thazhemadam <anant.thazhemadam@gmail.com>